requirements.txt
python-dotenv
langchain-community
langchain-groq